Conflicts:
	public/include/TXMPMeta.hpp
	public/include/XMP_Const.h
	samples/source/XMPScanner.cpp
	source/XMPCore/XMPUtils.cpp
	source/XMPFiles/FileHandlers/TIFF_Handler.cpp
	source/XMPFiles/FormatSupport/ReconcileTIFF.cpp
	source/XMPFiles/FormatSupport/Reconcile_Impl.cpp
	source/XMPFiles/FormatSupport/TIFF_FileWriter.cpp
	source/XMPFiles/FormatSupport/TIFF_MemoryReader.cpp
	source/XMPFiles/XMPFiles.cpp
	source/XMPFiles/XMPFiles_Impl.cpp
	source/common/EndianUtils.hpp

+++ source/XMPFiles/FormatSupport/PNG_Support.cpp
@@ -6,7 +6,6 @@
 // NOTICE: Adobe permits you to use, modify, and distribute this file in accordance with the terms
 // of the Adobe license agreement accompanying it.
 // =================================================================================================
-
 #include "PNG_Support.hpp"
 #include <string.h>
 
@@ -115,7 +114,7 @@ namespace PNG_Support
 		// read first and following chunks
 		while ( ReadChunk ( fileRef, inOutChunkState, &name, &len, pos) ) {}
 	
-		return inOutChunkState.chunks.size();
+		return (long)inOutChunkState.chunks.size();
 
 	}
